requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-bootstrap-icons==0.8.3
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.1
django-location-field==2.7.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-flex-fields==1.0.2
idna==3.4
Markdown==3.4.3
mysqlclient==2.1.1
phonenumbers==8.13.9
Pillow==9.5.0
pipreqs==0.4.13
PyJWT==2.6.0
pytz==2023.3
reportlab==3.6.12
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
yarg==0.1.9
